+++ TablePro/Views/Results/Cells/DataGridCellContent.swift
@@ -20,7 +20,6 @@ struct DataGridCellContent {
 struct DataGridCellState {
     let visualState: RowVisualState
     let isFocused: Bool
-    let isInCellSelection: Bool
     let isEditable: Bool
     let isLargeDataset: Bool
     let row: Int